Refactor /pool mountpoint check
Virtual workernodes use xfs as the root filesystem whereas physical workernodes have a dedicated mountpoint for /pool. This refactor accomodates both situations as well as allowing for ext4 should this ever be required for /pool.

This change uses `df -T` to derive the target mount of /pool and inspects the xfs format. If the xfs filesystem isn't formatted with `ftype=1` then the script fatally errors.