feat: redirect traffic to webdav.echo.stfc.ac.uk to local gateway
The endpoint in question is used mostly for writing data to ECHO.
Local gateways are writable now, so it's preferable to use them instead
of external ones (to separate the traffic better and reduce the load
on external gws).

We also remove xrootd-gateway endpoint since it is no longer needed
(local gateways are not used for xrootd uploads due to possible race
conditions).